+++ docs-markdown/src/helper/common.ts
@@ -1,11 +1,36 @@
 "use-strict";
 
-import os = require("os");
-import path = require("path");
+import * as fs from "fs";
+import * as glob from "glob";
+import * as os from "os";
+import * as path from "path";
 import * as vscode from "vscode";
 import { output } from "../extension";
 import * as log from "./log";
 import { reporter } from "./telemetry";
+
+export function tryFindFile(rootPath: string, fileName: string) {
+    try {
+        const fullPath = path.resolve(rootPath, fileName);
+        const exists = fs.existsSync(fullPath);
+        if (exists) {
+            return fullPath;
+        } else {
+            const files = glob.sync(`**/${fileName}`, {
+                cwd: rootPath,
+            });
+
+            if (files && files.length === 1) {
+                return path.join(rootPath, files[0]);
+            }
+        }
+    } catch (error) {
+        postError(error.toString());
+    }
+
+    postWarning(`Unable to find a file named "${fileName}", recursively at root "${rootPath}".`);
+    return undefined;
+}
 
 /**
  * Provide current os platform
@@ -75,6 +100,10 @@ export function noActiveEditorMessage() {
     postWarning("No active editor. Abandoning command.");
 }
 
+export function unsupportedFileMessage(languageId: string) {
+    postWarning(`Command is not support for "${languageId}". Abandoning command.`);
+}
+
 export function GetEditorText(editor: vscode.TextEditor, senderName: string): string {
     let content = "";
     const emptyString = "";
@@ -133,7 +162,7 @@ export function hasValidWorkSpaceRootPath(senderName: string) {
  * If overwrite is set the content will replace current selection.
  * @param {vscode.TextEditor} editor - the active editor in vs code.
  * @param {string} senderName - the name of the function that is calling this function
- * which is used to provide tracibility in logging.
+ * which is used to provide traceability in logging.
  * @param {string} string - the content to insert.
  * @param {boolean} overwrite - if true replaces current selection.
  * @param {vscode.Range} selection - if null uses the current selection for the insert or update.
@@ -267,6 +296,10 @@ export function isMarkdownFileCheck(editor: vscode.TextEditor, languageId: boole
     }
 }
 
+export function isValidFileCheck(editor: vscode.TextEditor, languageIds: string[]) {
+    return languageIds.some((id) => editor.document.languageId === id);
+}
+
 /**
  * Telemetry or Trace Log Type
  */
@@ -305,7 +338,7 @@ export function checkExtension(extensionName: string, notInstalledMessage?: stri
  */
 export function showStatusMessage(message: string) {
     const { msTimeValue } = generateTimestamp();
-    output.appendLine(`[${msTimeValue}] - ` + message);
+    output.appendLine(`[${msTimeValue}] - ${message}`);
 }
 
 /**
@@ -346,13 +379,22 @@ export async function showWarningMessage(message: string) {
 
 export function matchAll(
     pattern: RegExp,
-    text: string
-): Array<RegExpMatchArray> {
+    text: string,
+): RegExpMatchArray[] {
     const out: RegExpMatchArray[] = [];
     pattern.lastIndex = 0;
-    let match: RegExpMatchArray | null;
-    while ((match = pattern.exec(text))) {
-        out.push(match);
+    let match: RegExpMatchArray | null = pattern.exec(text);
+    while (match) {
+        if (match) {
+            // This is necessary to avoid infinite loops with zero-width matches
+            if (match.index === pattern.lastIndex) {
+                pattern.lastIndex++;
+            }
+
+            out.push(match);
+        }
+
+        match = pattern.exec(text);
     }
     return out;
 }
@@ -360,14 +402,14 @@ export function matchAll(
 export function extractDocumentLink(
     document: vscode.TextDocument,
     link: string,
-    matchIndex: number | undefined
+    matchIndex: number | undefined,
 ): vscode.DocumentLink | undefined {
     const offset = (matchIndex || 0) + 8;
     const linkStart = document.positionAt(offset);
     const linkEnd = document.positionAt(offset + link.length);
-    const text = document.getText(new vscode.Range(linkStart, linkEnd))
+    const text = document.getText(new vscode.Range(linkStart, linkEnd));
     try {
-        const httpMatch = text.match(/^(http|https):\/\//)
+        const httpMatch = text.match(/^(http|https):\/\//);
         if (httpMatch) {
             const documentLink = new vscode.DocumentLink(
                 new vscode.Range(linkStart, linkEnd),
@@ -386,3 +428,7 @@ export function extractDocumentLink(
         return undefined;
     }
 }
+
+export const naturalLanguageCompare = (a: string, b: string) => {
+    return (!!a && !!b) ? a.localeCompare(b, undefined, { numeric: true, sensitivity: "base" }) : 0;
+};